Update edac-tests to use ras-mc-ctl instead of ./edac-ctl
All functionalities previously found on my test version of
edac-ctl is present on ras-mc-ctl. So, let's rename it.

The test code still tries to run edac-util. This tool,
which is part of edac-utils, use the edac error counters to
check the errors. For now, let's keep it, as it might be useful,
although this will likely be removed on future versions of this
testing script.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>