rasdaemon: ras-mc-ctl: Add handling of run-time errors reported when querying HiSilicon KunPeng9xx OEM errors
Add try-catch for handling the run-time errors reported when querying
the HiSilicon KunPeng9xx OEM errors.
Reason:
For example, when an error table is not present in the SQLite DB,
then the DBI report error "no such table" and the ras-mc-ctl would exit
without query and log the other error types.
In the rasdaemon, for the vendor specific errors, the error table in the
SQLite database would be created, if not already done, when the corresponding
error is reported for the first time.

Reported-by: Lei Feng <fenglei47@h-partners.com>
Signed-off-by: Shiju Jose <shiju.jose@huawei.com>