rasdaemon: Fix CPU isolate errors with limited CPUs at startup
Since the patch f1ea76375281 ("rasdaemon: Check CPUs online, not
configured"), get_num_cpu() returns the number of online CPUs instead of
the total number of CPUs. This causes ras_cpu_isolation_init() to create
an array "cpu_infos" with the number of online CPUs. When reading the
array "cpu_infos", the CPU ID corresponds to the array index. If the user
limits the number of CPUs at system startup using the boot parameter
"maxcpus=xx", and then starts "rasdaemon" and brings another CPU online,
the number of online CPUs will exceed the number of members in the array
"cpu_infos".

This results in out-of-bounds access to the "cpu_infos" array when
isolating the newly online CPU, leading to unexpected errors.

error logs:
  [root@localhost ~]# rasdaemon: 2 error in cpu core caught
  rasdaemon: The current cpu 200 has exceed the total number of cpu:41
  rasdaemon: arm_event store: 0x40002400aa08
  rasdaemon: register inserted at db

  [1]+ Segmentation fault     (core dumped) rasdaemon -r -f
  [root@localhost ~]#

To fix this, we need to pass the total number of CPUs
(_SC_NPROCESSORS_CONF) in the system to ras_cpu_isolation_init().

Fixes: f1ea76375281 ("rasdaemon: Check CPUs online, not configured")
Signed-off-by: Junhao He <hejunhao3@h-partners.com>
Signed-off-by: Shiju Jose <shiju.jose@huawei.com>