rasdaemon: ras-events: Don't hardcode long size
The long's size is hardcoded to 64 bits. It prevents the event parsing to
work on 32-bits architecture.

Use the KBUFFER_LSIZE_SAME_AS_HOST to set the long's size so
libtraceevent will automatically select the relevant size.
Also use KBUFFER_ENDIAN_SAME_AS_HOST to set the endianness for
consistency.

Signed-off-by: Bastien Curutchet <bastien.curutchet@bootlin.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>